requirements/requirements-linux-python3.txt
pyqt5==5.11.3
lxml==4.6.3
numpy
opencv-python
--extra-index-url https://pypi.aiviro.com
aiviro-core==2.6.1
